# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Universe Moon [Um] API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members:
    get:
      operationId: getMembers
      tags: [members]
      summary: Get all members
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Create member (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemberRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    put:
      operationId: updateMember
      tags: [members]
      summary: Update member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete member (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /members/{id}/kick:
    post:
      operationId: kickMember
      tags: [members]
      summary: Kick member (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required: [reason]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /ngl:
    get:
      operationId: getNglMessages
      tags: [ngl]
      summary: Get NGL messages
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NglMessage"
    post:
      operationId: sendNglMessage
      tags: [ngl]
      summary: Send anonymous NGL message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendNglRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NglMessage"
  /ngl/{id}:
    delete:
      operationId: deleteNglMessage
      tags: [ngl]
      summary: Delete NGL message (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /games/imposter:
    get:
      operationId: getImposterGame
      tags: [games]
      summary: Get current imposter game state
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameState"
    post:
      operationId: createImposterGame
      tags: [games]
      summary: Create/start imposter game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameState"
  /games/imposter/action:
    post:
      operationId: gameAction
      tags: [games]
      summary: Perform game action (vote, eliminate, etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameActionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameState"
...
  /reactions:
    post:
      operationId: addReaction
      tags: [reactions]
      summary: Add reaction to content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddReactionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /votes:
    get:
      operationId: getVotes
      tags: [votes]
      summary: Get votes for a topic
      parameters:
        - name: topic
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"
    post:
      operationId: castVote
      tags: [votes]
      summary: Cast a vote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CastVoteRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /votes/reset:
    post:
      operationId: resetVotes
      tags: [votes]
      summary: Reset votes (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                topic:
                  type: string
              required: [topic]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Member:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nickname:
          type: string
        role:
          type: string
        bio:
          type: string
        joinDate:
          type: string
        favoriteSong:
          type: string
        socialLinks:
          type: string
        avatarUrl:
          type: string
        isActive:
          type: boolean
        kickReason:
          type: string
        kickDate:
          type: string
      required: [id, name, nickname, isActive]
    CreateMemberRequest:
      type: object
      properties:
        name:
          type: string
        nickname:
          type: string
        role:
          type: string
        bio:
          type: string
        joinDate:
          type: string
        favoriteSong:
          type: string
        socialLinks:
          type: string
        avatarUrl:
          type: string
      required: [name, nickname]
    UpdateMemberRequest:
      type: object
      properties:
        name:
          type: string
        nickname:
          type: string
        role:
          type: string
        bio:
          type: string
        favoriteSong:
          type: string
        socialLinks:
          type: string
        avatarUrl:
          type: string
...
    NglMessage:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        recipient:
          type: string
        createdAt:
          type: string
        reactions:
          type: object
          additionalProperties:
            type: integer
      required: [id, content, createdAt]
    SendNglRequest:
      type: object
      properties:
        content:
          type: string
        recipient:
          type: string
      required: [content]
...
    GameState:
      type: object
      properties:
        id:
          type: integer
        gameType:
          type: string
        status:
          type: string
        players:
          type: array
          items:
            $ref: "#/components/schemas/GamePlayer"
        phase:
          type: string
        round:
          type: integer
        winner:
          type: string
        createdAt:
          type: string
      required: [id, gameType, status, players, phase]
    GamePlayer:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
        isAlive:
          type: boolean
        isRevealed:
          type: boolean
      required: [username, isAlive]
    CreateGameRequest:
      type: object
      properties:
        gameType:
          type: string
        players:
          type: array
          items:
            type: string
      required: [gameType, players]
    GameActionRequest:
      type: object
      properties:
        action:
          type: string
        targetUsername:
          type: string
        actorUsername:
          type: string
      required: [action, actorUsername]
...
    AddReactionRequest:
      type: object
      properties:
        contentType:
          type: string
        contentId:
          type: integer
        emoji:
          type: string
        username:
          type: string
      required: [contentType, contentId, emoji, username]
    VoteResult:
      type: object
      properties:
        topic:
          type: string
        votes:
          type: object
          additionalProperties:
            type: integer
        totalVotes:
          type: integer
        userVote:
          type: string
      required: [topic, totalVotes]
    CastVoteRequest:
      type: object
      properties:
        topic:
          type: string
        option:
          type: string
        username:
          type: string
      required: [topic, option, username]