# lib/api-zod/src/generated/types/createMemberRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Universe Moon [Um] API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMemberRequest {
  name: string;
  nickname: string;
  role?: string;
  bio?: string;
  joinDate?: string;
  favoriteSong?: string;
  socialLinks?: string;
  avatarUrl?: string;
}